requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
duckdb==1.2.1
    # via images (pyproject.toml)
google-api-core[grpc]==2.24.2
    # via google-cloud-artifact-registry
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-cloud-artifact-registry
google-cloud-artifact-registry==1.15.1
    # via images (pyproject.toml)
googleapis-common-protos[grpc]==1.69.1
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.1
    # via google-cloud-artifact-registry
grpcio==1.71.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
idna==3.10
    # via requests
numpy==2.2.3
    # via pandas
pandas==2.2.3
    # via images (pyproject.toml)
polars==1.24.0
    # via images (pyproject.toml)
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-artifact-registry
protobuf==5.29.3
    # via
    #   google-api-core
    #   google-cloud-artifact-registry
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.1
    # via pandas
requests==2.32.3
    # via google-api-core
rsa==4.9
    # via google-auth
six==1.17.0
    # via python-dateutil
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
